backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.9.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-genai>=1.0.0
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
Pillow
jinja2>=3.1.0
aiofiles>=23.0.0
gspread>=6.0.0
supabase>=2.0.0
